crc: Changed crc size from 16 to 32 bits.
Make change and see if any test fails. If so - fix failing tests. In our
case no tests are failing due to earlier refactor and using new type for
crc.

If new feature changes behavior instead of data structures, the same
approach as in bugfixing should be used - write unit tests for new
functionalities, then update production code to pass these tests.

+++ src/comm/comm_master.h
@@ -6,7 +6,7 @@
 
 typedef uint16_t comm_addr_t;
 typedef uint16_t comm_seq_t;
-typedef uint16_t comm_crc_t;
+typedef uint32_t comm_crc_t;
 
 enum
 {
